requirements.txt
pdf2image
pytesseract
pillow
poppler
tesseract
python-docx
nltk
pandas
numpy